+++ src/Aardvark.Base/Math/Colors/Color.cs
@@ -1514,6 +1514,7 @@ namespace Aardvark.Base
 
         public static readonly PixFormat ByteBW = new PixFormat(typeof(byte), Col.Format.BW);
         public static readonly PixFormat ByteGray = new PixFormat(typeof(byte), Col.Format.Gray);
+        public static readonly PixFormat ByteGrayAlpha = new PixFormat(typeof(byte), Col.Format.GrayAlpha);
         public static readonly PixFormat ByteRG = new PixFormat(typeof(byte), Col.Format.RG);
         public static readonly PixFormat ByteBGR = new PixFormat(typeof(byte), Col.Format.BGR);
         public static readonly PixFormat ByteRGB = new PixFormat(typeof(byte), Col.Format.RGB);
@@ -1524,6 +1525,7 @@ namespace Aardvark.Base
 
         public static readonly PixFormat SByteBW = new PixFormat(typeof(sbyte), Col.Format.BW);
         public static readonly PixFormat SByteGray = new PixFormat(typeof(sbyte), Col.Format.Gray);
+        public static readonly PixFormat SByteGrayAlpha = new PixFormat(typeof(sbyte), Col.Format.GrayAlpha);
         public static readonly PixFormat SByteRG = new PixFormat(typeof(sbyte), Col.Format.RG);
         public static readonly PixFormat SByteBGR = new PixFormat(typeof(sbyte), Col.Format.BGR);
         public static readonly PixFormat SByteRGB = new PixFormat(typeof(sbyte), Col.Format.RGB);
@@ -1533,6 +1535,7 @@ namespace Aardvark.Base
         public static readonly PixFormat SByteRGBP = new PixFormat(typeof(sbyte), Col.Format.RGBP);
 
         public static readonly PixFormat UShortGray = new PixFormat(typeof(ushort), Col.Format.Gray);
+        public static readonly PixFormat UShortGrayAlpha = new PixFormat(typeof(ushort), Col.Format.GrayAlpha);
         public static readonly PixFormat UShortRG = new PixFormat(typeof(ushort), Col.Format.RG);
         public static readonly PixFormat UShortBGR = new PixFormat(typeof(ushort), Col.Format.BGR);
         public static readonly PixFormat UShortRGB = new PixFormat(typeof(ushort), Col.Format.RGB);
@@ -1542,6 +1545,7 @@ namespace Aardvark.Base
         public static readonly PixFormat UShortRGBP = new PixFormat(typeof(ushort), Col.Format.RGBP);
 
         public static readonly PixFormat ShortGray = new PixFormat(typeof(short), Col.Format.Gray);
+        public static readonly PixFormat ShortGrayAlpha = new PixFormat(typeof(short), Col.Format.GrayAlpha);
         public static readonly PixFormat ShortRG = new PixFormat(typeof(short), Col.Format.RG);
         public static readonly PixFormat ShortBGR = new PixFormat(typeof(short), Col.Format.BGR);
         public static readonly PixFormat ShortRGB = new PixFormat(typeof(short), Col.Format.RGB);
@@ -1551,6 +1555,7 @@ namespace Aardvark.Base
         public static readonly PixFormat ShortRGBP = new PixFormat(typeof(short), Col.Format.RGBP);
 
         public static readonly PixFormat UIntGray = new PixFormat(typeof(uint), Col.Format.Gray);
+        public static readonly PixFormat UIntGrayAlpha = new PixFormat(typeof(uint), Col.Format.GrayAlpha);
         public static readonly PixFormat UIntRG = new PixFormat(typeof(uint), Col.Format.RG);
         public static readonly PixFormat UIntBGR = new PixFormat(typeof(uint), Col.Format.BGR);
         public static readonly PixFormat UIntRGB = new PixFormat(typeof(uint), Col.Format.RGB);
@@ -1560,6 +1565,7 @@ namespace Aardvark.Base
         public static readonly PixFormat UIntRGBP = new PixFormat(typeof(uint), Col.Format.RGBP);
 
         public static readonly PixFormat IntGray = new PixFormat(typeof(int), Col.Format.Gray);
+        public static readonly PixFormat IntGrayAlpha = new PixFormat(typeof(int), Col.Format.GrayAlpha);
         public static readonly PixFormat IntRG = new PixFormat(typeof(int), Col.Format.RG);
         public static readonly PixFormat IntBGR = new PixFormat(typeof(int), Col.Format.BGR);
         public static readonly PixFormat IntRGB = new PixFormat(typeof(int), Col.Format.RGB);
@@ -1569,6 +1575,7 @@ namespace Aardvark.Base
         public static readonly PixFormat IntRGBP = new PixFormat(typeof(int), Col.Format.RGBP);
 
         public static readonly PixFormat HalfGray = new PixFormat(typeof(Half), Col.Format.Gray);
+        public static readonly PixFormat HalfGrayAlpha = new PixFormat(typeof(Half), Col.Format.GrayAlpha);
         public static readonly PixFormat HalfRG = new PixFormat(typeof(Half), Col.Format.RG);
         public static readonly PixFormat HalfBGR = new PixFormat(typeof(Half), Col.Format.BGR);
         public static readonly PixFormat HalfRGB = new PixFormat(typeof(Half), Col.Format.RGB);
@@ -1578,6 +1585,7 @@ namespace Aardvark.Base
         public static readonly PixFormat HalfRGBP = new PixFormat(typeof(Half), Col.Format.RGBP);
 
         public static readonly PixFormat FloatGray = new PixFormat(typeof(float), Col.Format.Gray);
+        public static readonly PixFormat FloatGrayAlpha = new PixFormat(typeof(float), Col.Format.GrayAlpha);
         public static readonly PixFormat FloatRG = new PixFormat(typeof(float), Col.Format.RG);
         public static readonly PixFormat FloatBGR = new PixFormat(typeof(float), Col.Format.BGR);
         public static readonly PixFormat FloatRGB = new PixFormat(typeof(float), Col.Format.RGB);
@@ -1587,6 +1595,7 @@ namespace Aardvark.Base
         public static readonly PixFormat FloatRGBP = new PixFormat(typeof(float), Col.Format.RGBP);
 
         public static readonly PixFormat DoubleGray = new PixFormat(typeof(double), Col.Format.Gray);
+        public static readonly PixFormat DoubleGrayAlpha = new PixFormat(typeof(double), Col.Format.GrayAlpha);
         public static readonly PixFormat DoubleRG = new PixFormat(typeof(double), Col.Format.RG);
         public static readonly PixFormat DoubleBGR = new PixFormat(typeof(double), Col.Format.BGR);
         public static readonly PixFormat DoubleRGB = new PixFormat(typeof(double), Col.Format.RGB);